+++ package.json
@@ -9,6 +9,7 @@
     "browser-sync": "^2.26.7",
     "gulp": "^4.0.2",
     "gulp-csso": "^4.0.1",
+    "gulp-htmlmin": "^5.0.1",
     "gulp-imagemin": "^6.2.0",
     "gulp-plumber": "1.2.x",
     "gulp-postcss": "8.0.x",
@@ -17,6 +18,7 @@
     "gulp-sass": "^4.0.2",
     "gulp-sourcemaps": "2.6.x",
     "gulp-svgstore": "^7.0.1",
+    "gulp-uglify": "^3.0.2",
     "gulp-webp": "^4.0.1",
     "posthtml-include": "^1.3.2",
     "stylelint": "10.1.x",
