+++ source/form.html
@@ -4,6 +4,8 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <title>HTML Academy: Седона</title>
+    <link rel="preload" href="fonts/ptsans.woff2" as="font" crossorigin="anonymous">
+    <link rel="preload" href="fonts/ptsansbold.woff2" as="font" crossorigin="anonymous">
     <link rel="stylesheet" href="css/style.min.css">
   </head>
   <body>
@@ -187,7 +189,6 @@
         </footer>
         </div>
     </div>
-    <script src="js/script.js"></script>
-    <script src="js/form.js"></script>
+    <script src="js/script.min.js"></script>
   </body>
 </html>
